backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.7.1
openai==1.58.1
google-generativeai==0.8.4
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
